Drop species with given patterns in name

Drop species with given patterns in name

Usage

DropPatterns(
  data,
  patterns = c("sp\\.", "\\/", "Domestic type", "hybrid"),
  colname = "comName"
)

Arguments

data

A data.frame with observations, likely the object from a call to RecentNearby.

patterns

Character vector of patterns in colname column to omit; interprets regular expressions.

colname

Character vector indicating column with names to search for patterns.

Value

Copy of data with any rows that has values in colname that match values in patterns.

Details

This utility function provides a means of omitting observations of "other taxa" such as domestics, hybrids, and "sp." observations. If patterns = NULL, no rows will be dropped.

Examples

df <- data.frame(comName = c("Mallard", "Mallard x Mexican Duck hybrid", "Verdin"),
                 date = c("2021-01-09", "2021-01-09", "2021-01-09"))
df <- DropPatterns(data = df)